HP-843 Move common decorators to open_city_profile module
That's a more appropriate place for common stuff.
